fix(rbac): move admin creation to its own file
